+++ packages/node/src/init.ts
@@ -2,8 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { NestFactory } from '@nestjs/core';
-import { findAvailablePort } from '@subql/common';
-import { getLogger, NestLogger } from '@subql/node-core';
+import { getLogger, getValidPort, NestLogger } from '@subql/node-core';
 import { AppModule } from './app.module';
 import { ApiService } from './indexer/api.service';
 import { FetchService } from './indexer/fetch.service';
@@ -13,28 +12,12 @@ const pjson = require('../package.json');
 
 const { argv } = yargsOptions;
 
-const DEFAULT_PORT = 3000;
 const logger = getLogger('subql-node');
 
 export async function bootstrap(): Promise<void> {
   logger.info(`Current ${pjson.name} version is ${pjson.version}`);
   const debug = argv.debug;
-
-  const validate = (x: any) => {
-    const p = parseInt(x);
-    return isNaN(p) ? null : p;
-  };
-
-  const port = validate(argv.port) ?? (await findAvailablePort(DEFAULT_PORT));
-  if (!port) {
-    logger.error(
-      `Unable to find available port (tried ports in range (${port}..${
-        port + 10
-      })). Try setting a free port manually by setting the --port flag`,
-    );
-    process.exit(1);
-  }
-
+  const port = await getValidPort(argv.port);
   if (argv.unsafe) {
     logger.warn(
       'UNSAFE MODE IS ENABLED. This is not recommended for most projects and will not be supported by our hosted service',
